backend/requirements.txt
<<<<<<< HEAD
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP
httpx==0.25.2
requests==2.31.0

# Async
aiofiles==23.2.1

# Web Scraping & Automation
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Vision & LLM
openai==1.3.9
easyocr==1.7.0

# Automation
pyautogui==0.9.53
pillow==10.1.0

# Task Scheduling
apscheduler==3.10.4

# Google Maps
google-maps-services==1.6.3

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Dev
black==23.12.0
flake8==6.1.0
isort==5.13.2

=======
fastapi
uvicorn
uvloop
sqlalchemy
pydantic
pydantic-settings
python-dotenv
requests
playwright
alembic
psycopg[binary]
pytest
pytest-asyncio
httpx
pyautogui
aiofiles
>>>>>>> 97953c3 (Initial commit from Specify template)
